feat: se construye el servicio de registro de paciente y usuauro desde auth hasta apaciente

+++ auth-service/src/main/java/com/pe/den/authservice/repository/RolRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface RolRepository
-        extends JpaRepository<Rol, Long> {
+        extends JpaRepository<Rol, Integer> {
 
     Optional<Rol> findByNombre(String nombre);
 }
